+++ src/main/java/com/nwidart/Application.java
@@ -1,12 +1,14 @@
 package com.nwidart;
 
 import com.vaadin.flow.component.page.AppShellConfigurator;
+import com.vaadin.flow.component.page.Push;
 import com.vaadin.flow.theme.Theme;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
 @Theme("default")
+@Push
 public class Application implements AppShellConfigurator {
 
     public static void main(String[] args) {